+++ src/utils/uri.ts
@@ -13,7 +13,9 @@ export function pathToUri(path: string): vscode.Uri | undefined {
   return vscode.Uri.joinPath(workspaceFolder.uri, path.slice(workspaceFolder.name.length + 1))
 }
 
-export function uriToPath(uri: vscode.Uri): string | undefined {
+export function uriToPath(uri: vscode.Uri | undefined): string | undefined {
+  if (!uri) return undefined
+
   const exactWorkspaceFolder = vscode.workspace.workspaceFolders?.find(workspaceFolder => workspaceFolder.uri.path === uri.path)
   if (exactWorkspaceFolder) {
     return exactWorkspaceFolder.name