+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Scoped Paths
 
 - **Added experimental "Hide Workspace Folders" setting** to completely hide workspace folders outside of the current scope
+- **Added dynamic scopes for Workspace Folders to quickly switch between them**
 - **Added configurable "Highlight status bar items when scope is enabled" setting** (enabled by default)
 - Related file labels are now smartly shortened when exceeding certain length
   > Configurable with `streamline.relatedFiles.maxLabelLength` and `streamline.relatedFiles.collapsedIndicator`