+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 0.18
+
+### Scoped Paths
+
+- **Added "Quick Scope" functionality** to temporarily scope into a workspace folder or any other path
+  > Available as a "Quick Scope into Selected" command and as dynamic scopes in "Change Current Scope..." command
+- **Added configurable "Highlight status bar items when scope is enabled" setting** (enabled by default)
+- Related file labels are now smartly shortened when exceeding certain length
+  > Configurable with `streamline.relatedFiles.maxLabelLength` and `streamline.relatedFiles.collapsedIndicator`
+
+### Tab History
+
+- **Added "Quick Open from Tab History" command**
+
 ## 0.17
 
 - Significantly optimized extension startup time and overall performance